safari-534.53-branch/LayoutTests/media/video-currentTime-delay.html

<html>
<body>

    <video controls></video>
    
    <p>Test a delay in playing the movie results in a canPlay event.</p>
    
    <script src=media-file.js></script>
    <script src=video-test.js></script>
    <script>
    waitForEvent('canplaythrough',  
    function () { 
        waitForEventAndEnd('ended');
    
        testExpected("video.currentTime", 0);
        run("video.currentTime = video.duration - 0.2"); 
    });

    waitForEvent('seeked',
    function () {        
        setTimeout(function() {
            run("video.play()"); 
        }, 400);
    
        var timeout = 2000;
        setTimeout(function () { logResult(false, "No 'ended' event in " + timeout/1000  + " seconds"); }, timeout);

    });
    
    video.src = findMediaFile("video", "content/test");
    </script>

</body>
</html>
